03_source/dev/miniDev-fix-qdrant-all-collections/python/requirements.txt
# Основные зависимости для Qdrant Manager
requests>=2.25.0
PyYAML>=5.4.0
qdrant-client>=1.6.0

# Опциональные зависимости для тестирования
pytest>=7.0.0
pytest-cov>=4.0.0
coverage>=7.0.0
mock>=4.0.3